refactor: move upload.py from dataset/ to core/upload_layout.py (Phase 2 §5.7)
dataset/upload.py was lying about its home — zero in-package
consumers; every caller lived in core/, asset/, schema/, or
execution/. Phase-2 dataset audit recommended ``execution/``
but pre-existing import-cycle structure (dataset/split.py:86
imports back into execution/) makes ``core/`` the safer landing
for a file that has callers across four packages.

Move + import rewrites:

* git mv src/deriva_ml/dataset/upload.py
        src/deriva_ml/core/upload_layout.py
* Updated 8 src callers and 1 test caller via
  deriva_ml.dataset.upload → deriva_ml.core.upload_layout.

Verified: ``import deriva_ml`` and
``from deriva_ml.core.upload_layout import (upload_root,
NULL_SENTINEL, asset_table_upload_spec,
bulk_upload_configuration)`` resolve cleanly.

The audit-suggested ``execution/upload_layout`` location induces
an import cycle: schema/annotations.py imports
``bulk_upload_configuration``, which would trigger
execution/__init__.py loading, which imports
ExecutionConfiguration, which transitively triggers
dataset/__init__.py, which loads dataset/split.py, which imports
back into execution/ before its __init__ finishes — circular.
``core/`` doesn't have that hazard because the four caller
packages (core, asset, schema, execution) are peers or
descendants of core.

LoC: ±0 (relocation).
